Refactored folder
