+++ src/projects/new/projectState.tsx
@@ -1,0 +1,7 @@
+interface projectState {
+    projectName: string,
+    baudRate: number,
+    analystInitials: string,
+}
+
+export default projectState
